git-svn-id: https://cubaja.googlecode.com/svn/trunk@327 dd2ac39a-1c43-11de-a027-6f5b8e30e268

+++ src/de/ufinke/cubaja/sql/SequenceConfig.java
@@ -12,7 +12,7 @@ public class SequenceConfig {
   private DatabaseConfig database;
   private String tableName;
   private String seqName;
-  private short blockSize;
+  private int blockSize;
 
   public SequenceConfig() {
 
@@ -72,7 +72,7 @@ public class SequenceConfig {
     this.seqName = seqName;
   }
 
-  public short getBlockSize() {
+  public int getBlockSize() {
 
     if (blockSize == 0) {
       blockSize = 10000;
@@ -81,7 +81,7 @@ public class SequenceConfig {
     return blockSize;
   }
 
-  public void setBlockSize(short blockSize) {
+  public void setBlockSize(int blockSize) {
 
     this.blockSize = blockSize;
   }
